Move canvas stuff into a separate directory to start decoupling it from GoogleRealtimeService and GraphModule
