+++ src/app/canvas/canvas/canvas.component.css
