+++ src/app/canvas/card/card.component.css
