+++ src/app/canvas/card/card.component.html
